# assets/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "# ChatPuppy Api Document\n\n## Auth\n- Add headers `Authorization Bearer {token}`\n- Add headers `X-Member-Address ${address}`\n\n## Signature\n1. Just post request\n2. Encrypt raw post body string to `MD5`, getting string `bodyMD5`\n3. Concatenate `timestamp` (from 1970, 10 digits) string end of step 2, getting string `s = bodyMD5 + timestamp`\n4. Sign `s` getting `signature` string\n5. Add headers `X-Signature ${signature}`\n6. Add headers `X-Timestamp ${timestamp}`\n\n## WebSocket\n\u003e `/message`\n\n### Operates\n- 1: register\n- 2: set group key\n- 3: new group message\n",
        "title": "ChatPuppy Api Doc",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/member/update": {
            "post": {
                "description": "need signature verify",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Member"
                ],
                "summary": "Update member's info",
                "operationId": "MemberUpdate",
                "parameters": [
                    {
                        "description": "Member's info",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MemberUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        },
...
        "/message": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "List message",
                "operationId": "MessageList",
                "parameters": [
                    {
                        "type": "string",
                        "name": "groupId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "last time",
                        "name": "time",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/model.Message"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Create message",
                "operationId": "MessageCreate",
                "parameters": [
                    {
                        "description": "message detail",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MessageCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/app.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.ChatMessageCreateRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/message/read": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Message"
                ],
                "summary": "Mask message read",
                "operationId": "MessageRead",
                "parameters": [
                    {
                        "description": "group id",
                        "name": "groupId",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response success",
                        "schema": {
                            "$ref": "#/definitions/app.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "app.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "model.ChatMessageCreateRes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "model.Member": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "avatar": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "model.MemberUpdateReq": {
            "type": "object",
            "properties": {
                "intro": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "model.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "member": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Member"
                        }
                    ]
                },
                "quote": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Message"
                        }
                    ]
                }
            }
        },
        "model.MessageCreateReq": {
            "type": "object",
            "required": [
                "content",
                "groupId",
                "memberId"
            ],
            "properties": {
                "content": {
                    "description": "encrypted message",
                    "type": "string"
                },
                "groupId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "memberId": {
                    "type": "string"
                },
                "parentId": {
                    "description": "quote message id",
                    "type": "string"
                }
            }
        },
...
    }
}